refactor(sync): split S3 sync helpers
The S3 session-source work accumulated object metadata, session processing, and Claude sidecar hydration in one file while the review cycle was still active. Splitting those responsibilities makes the S3 path easier to audit and keeps future fixes localized without changing behavior.
